Лабораторная 1б на неблокирующих сокетах
